[ATOM SGLang] SGL plugin Attention Refractory (#863)
* add work log

* [ATOM-SGL][Attn refrac] Separate model-specific MLA from SGL full attention backend

* remove work log

* [ATOM-SGL][Attn refrac] Route DeepSeek MLA through an SGLang wrapper
Move the SGLang DeepSeek MLA runtime entry from legacy forward glue into
SGLangDeepseekMLAAttention while keeping RadixAttention and the full-attention
backend as the host/backend layers. Shrink deepseek_mla_forward.py into a
helper module and clarify absorbed vs non-absorbed path naming.

* [ATOM SGL] runtime extraction

* [ATOM-SGL][Runtime] Introduce model adapter specs


* [ATOM-SGL][Runtime] Keep custom wrappers out of generated entries


* [ATOM-SGL][Attn refrac] Split full attention backend helpers


* [ATOM-SGL][Attn refrac] Format refactored attention files


* [ATOM-SGL][Attn refrac] Fix ruff findings in refactored attention code


* [ATOM-SGL][Attn refrac] Avoid DeepSeek MLA wrapper module cycle


* fix rebase issue

* precheckin

* prepare for sglang only

* import error meet in qwen3.5

* qwen3.5 acc fix

* [Fix] Limit static FP4 linear kv_b_proj post-processing

---------
