+++ atom/plugin/sglang/__init__.py
@@ -1,0 +1,3 @@
+from atom.plugin.sglang.prepare import prepare_model_for_sglang
+
+__all__ = ["prepare_model_for_sglang"]
