+++ atom/plugin/sglang/models/qwen3_5.py
@@ -37,7 +37,7 @@ from atom.models.utils import IntermediateTensors
 from atom.plugin.sglang.attention_backend.attention_gdn import (
     SGLangGDNForwardContext,
 )
-from atom.plugin.sglang.models.base_model_wrapper import (
+from atom.plugin.sglang.runtime import (
     SGLangForwardBatchMetadata,
 )
 
@@ -202,13 +202,13 @@ def _get_qwen35_language_model_stack_cls(
             prefix: str = "",
         ) -> None:
             del prefix
-            import atom
+            from atom.plugin.sglang.prepare import prepare_model
 
             nn.Module.__init__(self)
             root_config = type(self)._pending_vlm_root_config
             if root_config is None:
                 root_config = config
-            atom_lm = atom.prepare_model(config=root_config, engine="sglang")
+            atom_lm = prepare_model(config=root_config)
             if atom_lm is None:
                 arch = getattr(root_config, "architectures", ["unknown"])[0]
                 raise ValueError(f"ATOM failed to build language model for {arch}")
@@ -448,5 +448,5 @@ class Qwen3_5MoeForConditionalGeneration(
 
 # SGLang discovers these multimodal wrappers from this module's `EntryClass`.
 # They are not covered by `base_model_wrapper.py`, whose generated entries only
-# handle the plain causal-LM architectures in `_MODEL_ARCH_SPECS`.
+# handle the plain causal-LM architectures in `MODEL_ARCH_SPECS`.
 EntryClass = [Qwen3_5ForConditionalGeneration, Qwen3_5MoeForConditionalGeneration]
