+++ atom/plugin/sglang/runtime/model_arch.py
@@ -1,0 +1,65 @@
+"""SGLang plugin model adapter registry."""
+
+from __future__ import annotations
+
+from dataclasses import dataclass
+from typing import Any, Callable, Optional
+
+
+@dataclass(frozen=True)
+class SGLangModelAdapterSpec:
+    """Adapter hooks for one SGLang plugin model architecture.
+
+    The first version keeps the existing runtime flags while adding function
+    hooks for config preparation and install-time model adaptation. This avoids
+    growing a long list of booleans in the generic wrapper as new models arrive.
+    """
+
+    wrapper_binds_gdn_context: bool = False
+    prepare_config: Optional[Callable[[Any, str], None]] = None
+    install_adapters: Optional[Callable[[Any], None]] = None
+
+
+def _prepare_qwen35_config(atom_config: Any, model_arch: str) -> None:
+    from atom.plugin.sglang.models.qwen3_5 import apply_prepare_model_adaptations
+
+    apply_prepare_model_adaptations(atom_config, model_arch)
+
+
+def _install_deepseek_mla_adapters(model: Any) -> None:
+    from atom.plugin.sglang.models.deepseek_mla import setup_deepseek_for_sglang
+
+    setup_deepseek_for_sglang(model)
+
+
+MODEL_ADAPTER_SPECS = {
+    "DeepseekV3ForCausalLM": SGLangModelAdapterSpec(
+        install_adapters=_install_deepseek_mla_adapters,
+    ),
+    "Qwen3MoeForCausalLM": SGLangModelAdapterSpec(),
+    "Qwen3NextForCausalLM": SGLangModelAdapterSpec(
+        wrapper_binds_gdn_context=True,
+    ),
+    "Qwen3_5ForConditionalGeneration": SGLangModelAdapterSpec(
+        prepare_config=_prepare_qwen35_config,
+    ),
+    "Qwen3_5MoeForConditionalGeneration": SGLangModelAdapterSpec(
+        prepare_config=_prepare_qwen35_config,
+    ),
+}
+
+# Architectures whose SGLang EntryClass is generated by base_model_wrapper.
+# Custom outer-wrapper modules, such as Qwen3.5 multimodal wrappers, keep their
+# own EntryClass and should not appear here or SGLang will see duplicate classes.
+MODEL_ARCH_SPECS = {
+    key: MODEL_ADAPTER_SPECS[key]
+    for key in (
+        "DeepseekV3ForCausalLM",
+        "Qwen3MoeForCausalLM",
+        "Qwen3NextForCausalLM",
+    )
+}
+
+
+def get_model_arch_spec(model_arch: str) -> SGLangModelAdapterSpec:
+    return MODEL_ADAPTER_SPECS.get(model_arch, SGLangModelAdapterSpec())
